refactor for more modular design
