+++ components/InputComponent.tsx
