+++ components/Mobile.tsx
@@ -1,5 +1,5 @@
 import Image from 'next/image';
-import logo from '../../public/unisg_logo.png';
+import logo from '../public/unisg_logo.png';
 
 export const MobileNotSupported: React.FC = () => {
   return (
