+++ lib/services/chatCompletionWithTools.tsx
@@ -2,10 +2,9 @@ import { ChatOpenAI } from '@langchain/openai';
 import { ChatGroq } from '@langchain/groq';
 
 // tools
-import { getCurrentWeather } from './tools/getWeather';
-import { searchSong } from './tools/getSpotify';
-import { getTime } from './tools/getTime';
-import { config } from '../config';
+import { getCurrentWeather } from '../tools/getWeather';
+import { searchSong } from '../tools/getSpotify';
+import { getTime } from '../tools/getTime';
 
 const tool_calls = new ChatOpenAI({
   model: 'gpt-3.5-turbo',
