+++ lib/services/generateAgentCompletion.tsx
@@ -3,7 +3,7 @@ import { ChatPromptTemplate, MessagesPlaceholder } from '@langchain/core/prompts
 import { ChatOpenAI } from '@langchain/openai';
 import { END, MemorySaver, StateGraph, StateGraphArgs, START } from '@langchain/langgraph';
 import { AIMessage, BaseMessage, HumanMessage } from '@langchain/core/messages';
-import { saveJson, getJson } from './mongoUtil';
+import { saveJson, getJson } from '../utils/mongoUtil';
 import { traceable } from 'langsmith/traceable';
 
 // Define the state interface which includes both messages and responses
