+++ lib/services/generateChainCompletion.tsx
