+++ lib/services/generateChatCompletion.tsx
