+++ lib/services/processImage.tsx
