+++ tsconfig.json
@@ -23,6 +23,8 @@
   },
   "include": [
     "next-env.d.ts",
+    "*.ts",
+    "*.tsx",
     "**/*.ts",
     "**/*.tsx",
     ".next/types/**/*.ts",
